Refactored code to address PR review comments
