--- app/javascript/src/components/Dashboard/Contacts/utils.js
@@ -1,5 +1,0 @@
-export const generateContactsData = (contacts, multiplier = 2) =>
-  Array(multiplier)
-    .fill(contacts)
-    .flat()
-    .map((contact, index) => ({ ...contact, id: index }));
